Update crunchyroll.en.bot.py
Change to PT time zone + additionally to UTC.

In addition, it now writes the ID to the list of entries already read on each run. Previously, the ID was only written to the file after the loop was completed.

This measure was actually intended to reduce the number of writes to the hard disc. However, if an error occurs during the execution of the loop, the entries that have already been sent are sent again.